requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.1
django-cors-headers==2.1.0
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
et-xmlfile==1.0.1
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
openpyxl==2.5.0
pkg-resources==0.0.0
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
uritemplate==3.0.0
urllib3==1.22
